feat: update searchResult (#221) ootd 검색 API 위치 변경
