+++ components/Domain/Search/SearchResult/ClosetCloth/index.tsx
@@ -16,16 +16,12 @@ import SubHead from '../SubHead';
 import { CategoryListType } from '@/components/Domain/AddCloth/ClothCategoryModal';
 import { BrandType } from '@/components/BrandList/Brand';
 import Spinner from '@/components/Spinner';
+import { OOTDApi } from '@/apis/domain/OOTD/OOTDApi';
+import useInfiniteScroll from '@/hooks/useInfiniteScroll';
+import useEffectAfterMount from '@/hooks/useEffectAfterMount';
 
 interface ClosetClothProps {
-  filter: FilterData;
-  setFilter: Dispatch<SetStateAction<FilterData>>;
-  sortStandard: string;
-  setSortStandard: Dispatch<SetStateAction<string>>;
-  OOTDList: OOTDListType[];
-  OOTDIsLoading: Boolean;
-  OOTDRef: MutableRefObject<any>;
-  OOTDHasNextPage: Boolean;
+  keywordsValue: string;
 }
 
 export type OOTDListType = {
@@ -45,25 +41,24 @@ export interface FilterData {
   gender: GenderTypes;
 }
 
-export default function ClosetCloth({
-  OOTDList,
-  filter,
-  setFilter,
-  sortStandard,
-  setSortStandard,
-  OOTDIsLoading,
-  OOTDRef,
-  OOTDHasNextPage,
-}: ClosetClothProps) {
+export default function ClosetCloth({ keywordsValue }: ClosetClothProps) {
   const router = useRouter();
 
   const [filterModalIsOpen, setFilterModalIsOpen] = useState<Boolean>(false);
+  const [filter, setFilter] = useState<FilterData>({
+    category: null,
+    color: null,
+    brand: null,
+    gender: {
+      man: false,
+      woman: false,
+    },
+  });
 
-  console.log(OOTDList);
+  const [sortStandard, setSortStandard] = useState<string>('LATEST');
+  const [OOTDList, setOOTDList] = useState<OOTDListType[]>([]);
 
-  useEffect(() => {
-    console.log(filter);
-  }, [filter]);
+  const { getSearchOOTD } = OOTDApi();
 
   const onClickImageList = (index: number) => {
     router.push(`/ootd/${index}`);
@@ -84,6 +79,79 @@ export default function ClosetCloth({
       },
     });
   };
+
+  const [genderData, setGenderData] = useState<string>('');
+
+  useEffect(() => {
+    if (filter.gender.man && filter.gender.woman) {
+      setGenderData('');
+    } else if (filter.gender.man) {
+      setGenderData('MALE');
+    } else if (filter.gender.woman) {
+      setGenderData('FEMALE');
+    } else {
+      setGenderData('');
+    }
+  }, [filter.gender]);
+
+  const fetchOOTDDataFunction = async (ootdPage: number, ootdSize: number) => {
+    if (!router.isReady) return;
+
+    const data = await getSearchOOTD({
+      searchText: keywordsValue,
+      categoryIds: filter.category?.map((item) => {
+        if (item.state) {
+          return item.id;
+        }
+        return item.detailCategories![0].id;
+      }),
+      colorIds: filter.color?.map((item) => item.id),
+      brandIds: filter.brand?.map((item) => item.id),
+      writerGender: genderData,
+      sortCriteria: sortStandard,
+      page: ootdPage,
+      size: ootdSize,
+    });
+
+    return data;
+  };
+
+  const {
+    data: OOTDData,
+    isLoading: OOTDIsLoading,
+    containerRef: OOTDRef,
+    hasNextPage: OOTDHasNextPage,
+    reset: ootdReset,
+  } = useInfiniteScroll({
+    fetchDataFunction: fetchOOTDDataFunction,
+    size: 9,
+    initialData: [],
+  });
+
+  useEffect(() => {
+    setOOTDList(
+      OOTDData.map((item: any) => {
+        return {
+          id: item.id,
+          imageUrl: item.imageUrl,
+        };
+      })
+    );
+  }, [OOTDData]);
+
+  useEffectAfterMount(() => {
+    setOOTDList([]);
+    ootdReset();
+    setFilter({
+      category: null,
+      color: null,
+      brand: null,
+      gender: {
+        man: false,
+        woman: false,
+      },
+    });
+  }, [keywordsValue, sortStandard]);
 
   console.log(OOTDRef);
